Error Fixed and Listed Books Section design complited

+++ src/Utility/LocalStorageForWishList.js
@@ -1,3 +1,5 @@
+import { getStoredReadBooks } from "./LocalStorage";
+
 const getStoredWishListBooks = () => {
     const storedWishListBook = localStorage.getItem('wishListBook');
     if (storedWishListBook) {
@@ -8,8 +10,10 @@ const getStoredWishListBooks = () => {
 
 const saveWishListBook = id => {
     const storedWishListBook = getStoredWishListBooks();
+    const storedReadBooks = getStoredReadBooks();
+    const isMatch = storedReadBooks.find(Id=>Id===id);
     const exists = storedWishListBook.find(bookId => bookId === id);
-    if (!exists) {
+    if (!exists && !isMatch) {
         storedWishListBook.push(id);
         localStorage.setItem('wishListBook', JSON.stringify(storedWishListBook));
     }
